consolidate libutils
