+++ nd4j-backends/nd4j-api-parent/nd4j-native-api/src/main/java/org/nd4j/nativeblas/Nd4jBlas.java
@@ -4,7 +4,7 @@ package org.nd4j.nativeblas;
 import org.bytedeco.javacpp.Loader;
 import org.bytedeco.javacpp.Pointer;
 import org.bytedeco.javacpp.annotation.Platform;
-import org.nd4j.nativeblas.util.LibUtils;
+import org.nd4j.linalg.api.buffer.util.LibUtils;
 
 /**
  * CBlas bindings
