+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/jcuda/jcublas/JCublas.java
@@ -31,6 +31,7 @@ package jcuda.jcublas;
 import java.nio.*;
 import jcuda.*;
 import jcuda.runtime.*;
+import org.nd4j.linalg.api.buffer.util.LibUtils;
 
 /**
  * Java bindings for CUBLAS, the NVIDIA CUDA BLAS library.
