+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/jcuda/runtime/JCuda.java
@@ -28,6 +28,7 @@
 package jcuda.runtime;
 
 import jcuda.*;
+import org.nd4j.linalg.api.buffer.util.LibUtils;
 
 
 /**
