+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/jcusolver/JCusolver.java
@@ -28,6 +28,7 @@
 package jcusolver;
 
 import jcuda.*;
+import org.nd4j.linalg.api.buffer.util.LibUtils;
 
 /**
  * Java bindings for CUSOLVER, the NVIDIA CUDA solver library. <br />
